requirements.txt
# Core dependencies
boto3>=1.26.0
psycopg2-binary>=2.9.5
SQLAlchemy>=1.4.46
alembic>=1.9.1
pydantic>=1.10.4
python-dotenv>=0.21.1
click>=8.1.3
requests>=2.28.2
pyyaml>=6.0
inquirer==2.10.1
termcolor==2.3.0
rich==10.12.0
rio-ui>=0.11
openai>=1.16.2